Move common validation code to single place.
