+++ calico/common.py
@@ -43,6 +43,21 @@ FORMAT_STRING = '%(asctime)s [%(levelname)s][%(process)s/%(tid)d] %(name)s %(lin
 # string used by the logger.
 SYSLOG_FORMAT_STRING = '{excname}[%(process)s]: %(module)s@%(lineno)d %(message)s'
 
+# Valid keys for a rule JSON dict.
+KNOWN_RULE_KEYS = set([
+    "action",
+    "protocol",
+    "src_net",
+    "src_tag",
+    "src_ports",
+    "dst_net",
+    "dst_tag",
+    "dst_ports",
+    "icmp_type",
+    "icmp_code",
+    "ip_version",
+])
+
 tid_storage = gevent.local.local()
 tid_counter = itertools.count()
 # Ought to do itertools.count(start=1), but python 2.6 does not support it.
@@ -216,3 +231,216 @@ def complete_logging(logfile=None,
             root_logger.addHandler(file_handler)
 
     _log.info("Logging initialized")
+
+
+class ValidationFailed(Exception):
+    """
+    Class used for data validation exceptions.
+    """
+    pass
+
+
+def validate_endpoint(config, endpoint):
+    """
+    Ensures that the supplied endpoint is valid. Once this routine has returned
+    successfully, we know that all required fields are present and have valid
+    values.
+
+    :param config: configuration structure
+    :param endpoint: endpoint dictionary as read from etcd
+    :raises ValidationFailed
+    """
+    issues = []
+
+    if not isinstance(endpoint, dict):
+        raise ValidationFailed("Expected endpoint to be a dict.")
+
+    if "state" not in endpoint:
+        issues.append("Missing 'state' field.")
+    elif endpoint["state"] not in ("active", "inactive"):
+        issues.append("Expected 'state' to be one of active/inactive.")
+
+    for field in ["name", "mac", "profile_id"]:
+        if field not in endpoint:
+            issues.append("Missing '%s' field." % field)
+        elif not isinstance(endpoint[field], StringTypes):
+            issues.append("Expected '%s' to be a string; got %r." %
+                          (field, endpoint[field]))
+
+    if "name" in endpoint:
+        if not endpoint["name"].startswith(config.IFACE_PREFIX):
+            issues.append("Interface %r does not start with %r." %
+                          (endpoint["name"], config.IFACE_PREFIX))
+
+    for version in (4, 6):
+        nets = "ipv%d_nets" % version
+        if nets not in endpoint:
+            issues.append("Missing network %s." % nets)
+        else:
+            for ip in endpoint.get(nets, []):
+                if not validate_cidr(ip, version):
+                    issues.append("IP address %r is not a valid IPv%d CIDR." %
+                                  (ip, version))
+                    break
+
+        gw_key = "ipv%d_gateway" % version
+        try:
+            gw_str = endpoint[gw_key]
+            if gw_str is not None and not common.validate_ip_addr(gw_str,
+                                                                  version):
+                issues.append("%s is not a valid IPv%d gateway address." %
+                              (gw_key, version))
+        except KeyError:
+            pass
+
+    if issues:
+        raise ValidationFailed(" ".join(issues))
+
+def validate_rules(rules):
+    """
+    Ensures that the supplied rules are valid. Once this routine has returned
+    successfully, we know that all required fields are present and have valid
+    values.
+
+    :param rules: rules list as read from etcd
+    :raises ValidationFailed
+    """
+    issues = []
+
+    if not isinstance(rules, dict):
+        raise ValidationFailed("Expected rules to be a dict.")
+
+    for dirn in ("inbound_rules", "outbound_rules"):
+        if dirn not in rules:
+            issues.append("No %s in rules." % dirn)
+            continue
+
+        if not isinstance(rules[dirn], list):
+            issues.append("Expected rules[%s] to be a dict." % dirn)
+            continue
+
+        for rule in rules[dirn]:
+            # Absolutely all fields are optional, but some have valid and
+            # invalid values.
+            protocol = rule.get('protocol')
+            if (protocol is not None and
+                not protocol in [ "tcp", "udp", "icmp", "icmpv6" ]):
+                    issues.append("Invalid protocol in rule %s." % rule)
+
+            ip_version = rule.get('ip_version')
+            if (ip_version is not None and
+                not ip_version in [ 4, 6 ]):
+                # Bad IP version prevents further validation
+                issues.append("Invalid ip_version in rule %s." % rule)
+                continue
+
+            if ip_version == 4 and protocol == "icmpv6":
+                issues.append("Using icmpv6 with IPv4 in rule %s." % rule)
+            if ip_version == 6 and protocol == "icmp":
+                issues.append("Using icmp with IPv6 in rule %s." % rule)
+
+            # TODO: Validate that src_tag and dst_tag contain only valid characters.
+
+            for key in ("src_net", "dst_net"):
+                network = rule.get(key)
+                if (network is not None and
+                    not validate_cidr(rule[key], ip_version)):
+                    issues.append("Invalid CIDR (version %s) in rule %s." %
+                                  (ip_version, rule))
+
+            for key in ("src_ports", "dst_ports"):
+                ports = rule.get(key)
+                if (ports is not None and
+                    not isinstance(ports, list)):
+                    issues.append("Expected ports to be a list in rule %s."
+                                  % rule)
+                    continue
+
+                if ports is not None:
+                    for port in ports:
+                        error = validate_rule_port(port)
+                        if error:
+                            issues.append("Invalid port %s (%s) in rule %s." %
+                                          (port, error, rule))
+
+            action = rule.get('action')
+            if (action is not None and
+                    action not in ("allow", "deny")):
+                issues.append("Invalid action in rule %s." % rule)
+
+            icmp_type = rule.get('icmp_type')
+            if icmp_type is not None:
+                if not 0 <= icmp_type <= 255:
+                    issues.append("ICMP type is out of range.")
+            icmp_code = rule.get("icmp_code")
+            if icmp_code is not None:
+                if not 0 <= icmp_code <= 255:
+                    issues.append("ICMP code is out of range.")
+                if icmp_type is None:
+                    # TODO: ICMP code without ICMP type not supported by iptables
+                    # Firewall against that for now.
+                    issues.append("ICMP code specified without ICMP type.")
+
+            unknown_keys = set(rule.keys()) - KNOWN_RULE_KEYS
+            if unknown_keys:
+                issues.append("Rule contains unknown keys: %s." % unknown_keys)
+
+    if issues:
+        raise ValidationFailed(" ".join(issues))
+
+
+def validate_rule_port(port):
+    """
+    Validates that any value in a port list really is valid.
+    Valid values are an integer port, or a string range separated by a colon.
+
+    :param port: the port, which is validated for type
+    :returns: None or an error string if invalid
+    """
+    if isinstance(port, int):
+        if port < 1 or port > 65535:
+            return "integer out of range"
+        return None
+
+    # If not an integer, must be format N:M, i.e. a port range.
+    try:
+        fields = port.split(":")
+    except AttributeError:
+        return "neither integer nor string"
+
+    if not len(fields) == 2:
+        return "range unparseable"
+
+    try:
+        start = int(fields.pop(0))
+        end = int(fields.pop(0))
+    except ValueError:
+        return "range invalid"
+
+    if start >= end or start < 1 or end > 65535:
+        return "range invalid"
+
+    return None
+
+
+def validate_tags(tags):
+    """
+    Ensures that the supplied tags are valid. Once this routine has returned
+    successfully, we know that all required fields are present and have valid
+    values.
+
+    :param tags: tag set as read from etcd
+    :raises ValidationFailed
+    """
+    issues = []
+
+    if not isinstance(tags, list):
+        issues.append("Expected tags to be a list.")
+    else:
+        for tag in tags:
+            if not isinstance(tag, StringTypes):
+                issues.append("Expected tag '%s' to be a string." % tag)
+                break
+
+    if issues:
+        raise ValidationFailed(" ".join(issues))
