+++ calico/felix/fetcd.py
@@ -32,13 +32,13 @@ import urllib3.exceptions
 from urllib3.exceptions import ReadTimeoutError, ConnectTimeoutError
 
 from calico import common
+from calico.common import ValidationFailed, KNOWN_RULE_KEYS
 from calico.datamodel_v1 import (VERSION_DIR, READY_KEY, CONFIG_DIR,
                                  RULES_KEY_RE, TAGS_KEY_RE, ENDPOINT_KEY_RE,
                                  dir_for_per_host_config,
                                  get_profile_id_for_profile_dir, dir_for_host,
                                  PROFILE_DIR, HOST_DIR)
 from calico.felix.actor import Actor, actor_message
-from calico.felix.frules import KNOWN_RULE_KEYS
 
 _log = logging.getLogger(__name__)
 
@@ -52,9 +52,6 @@ PREFIXES_TO_RESYNC_ON_CHANGE = [
     PROFILE_DIR,
     HOST_DIR,
 ]
-
-class ValidationFailed(Exception):
-    pass
 
 
 class EtcdWatcher(Actor):
@@ -381,7 +378,7 @@ def parse_if_endpoint(config, etcd_node):
             hostname = m.group("hostname")
             endpoint = json_decoder.decode(etcd_node.value)
             try:
-                validate_endpoint(config, endpoint)
+                common.validate_endpoint(config, endpoint)
             except ValidationFailed as e:
                 _log.warning("Validation failed for endpoint %s, treating as "
                              "missing: %s", endpoint_id, e.message)
@@ -391,63 +388,6 @@ def parse_if_endpoint(config, etcd_node):
             _log.debug("Found endpoint : %s", endpoint)
         return endpoint_id, endpoint
     return None, None
-
-
-def validate_endpoint(config, endpoint):
-    """
-    Ensures that the supplied endpoint is valid. Once this routine has returned
-    successfully, we know that all required fields are present and have valid
-    values.
-
-    :param config: configuration structure
-    :param endpoint: endpoint dictionary as read from etcd
-    :raises ValidationFailed
-    """
-    issues = []
-
-    if not isinstance(endpoint, dict):
-        raise ValidationFailed("Expected endpoint to be a dict.")
-
-    if "state" not in endpoint:
-        issues.append("Missing 'state' field.")
-    elif endpoint["state"] not in ("active", "inactive"):
-        issues.append("Expected 'state' to be one of active/inactive.")
-
-    for field in ["name", "mac", "profile_id"]:
-        if field not in endpoint:
-            issues.append("Missing '%s' field." % field)
-        elif not isinstance(endpoint[field], StringTypes):
-            issues.append("Expected '%s' to be a string; got %r." %
-                          (field, endpoint[field]))
-
-    if "name" in endpoint:
-        if not endpoint["name"].startswith(config.IFACE_PREFIX):
-            issues.append("Interface %r does not start with %r." %
-                          (endpoint["name"], config.IFACE_PREFIX))
-
-    for version in (4, 6):
-        nets = "ipv%d_nets" % version
-        if nets not in endpoint:
-            issues.append("Missing network %s." % nets)
-        else:
-            for ip in endpoint.get(nets, []):
-                if not common.validate_cidr(ip, version):
-                    issues.append("IP address %r is not a valid IPv%d CIDR." %
-                                  (ip, version))
-                    break
-
-        gw_key = "ipv%d_gateway" % version
-        try:
-            gw_str = endpoint[gw_key]
-            if gw_str is not None and not common.validate_ip_addr(gw_str,
-                                                                  version):
-                issues.append("%s is not a valid IPv%d gateway address." %
-                              (gw_key, version))
-        except KeyError:
-            pass
-
-    if issues:
-        raise ValidationFailed(" ".join(issues))
 
 
 def parse_if_rules(etcd_node):
@@ -461,7 +401,7 @@ def parse_if_rules(etcd_node):
             rules = json_decoder.decode(etcd_node.value)
             rules["id"] = profile_id
             try:
-                validate_rules(rules)
+                common.validate_rules(rules)
             except ValidationFailed:
                 _log.exception("Validation failed for profile %s rules: %s",
                                profile_id, rules)
@@ -471,133 +411,6 @@ def parse_if_rules(etcd_node):
 
         return profile_id, rules
     return None, None
-
-
-def validate_rules(rules):
-    """
-    Ensures that the supplied rules are valid. Once this routine has returned
-    successfully, we know that all required fields are present and have valid
-    values.
-
-    :param rules: rules list as read from etcd
-    :raises ValidationFailed
-    """
-    issues = []
-
-    if not isinstance(rules, dict):
-        raise ValidationFailed("Expected rules to be a dict.")
-
-    for dirn in ("inbound_rules", "outbound_rules"):
-        if dirn not in rules:
-            issues.append("No %s in rules." % dirn)
-            continue
-
-        if not isinstance(rules[dirn], list):
-            issues.append("Expected rules[%s] to be a dict." % dirn)
-            continue
-
-        for rule in rules[dirn]:
-            # Absolutely all fields are optional, but some have valid and
-            # invalid values.
-            protocol = rule.get('protocol')
-            if (protocol is not None and
-                not protocol in [ "tcp", "udp", "icmp", "icmpv6" ]):
-                    issues.append("Invalid protocol in rule %s." % rule)
-
-            ip_version = rule.get('ip_version')
-            if (ip_version is not None and
-                not ip_version in [ 4, 6 ]):
-                # Bad IP version prevents further validation
-                issues.append("Invalid ip_version in rule %s." % rule)
-                continue
-
-            if ip_version == 4 and protocol == "icmpv6":
-                issues.append("Using icmpv6 with IPv4 in rule %s." % rule)
-            if ip_version == 6 and protocol == "icmp":
-                issues.append("Using icmp with IPv6 in rule %s." % rule)
-
-            # TODO: Validate that src_tag and dst_tag contain only valid characters.
-
-            for key in ("src_net", "dst_net"):
-                network = rule.get(key)
-                if (network is not None and
-                    not common.validate_cidr(rule[key], ip_version)):
-                    issues.append("Invalid CIDR (version %s) in rule %s." %
-                                  (ip_version, rule))
-
-            for key in ("src_ports", "dst_ports"):
-                ports = rule.get(key)
-                if (ports is not None and
-                    not isinstance(ports, list)):
-                    issues.append("Expected ports to be a list in rule %s."
-                                  % rule)
-                    continue
-
-                if ports is not None:
-                    for port in ports:
-                        error = validate_rule_port(port)
-                        if error:
-                            issues.append("Invalid port %s (%s) in rule %s." %
-                                          (port, error, rule))
-
-            action = rule.get('action')
-            if (action is not None and
-                    action not in ("allow", "deny")):
-                issues.append("Invalid action in rule %s." % rule)
-
-            icmp_type = rule.get('icmp_type')
-            if icmp_type is not None:
-                if not 0 <= icmp_type <= 255:
-                    issues.append("ICMP type is out of range.")
-            icmp_code = rule.get("icmp_code")
-            if icmp_code is not None:
-                if not 0 <= icmp_code <= 255:
-                    issues.append("ICMP code is out of range.")
-                if icmp_type is None:
-                    # TODO: ICMP code without ICMP type not supported by iptables
-                    # Firewall against that for now.
-                    issues.append("ICMP code specified without ICMP type.")
-
-            unknown_keys = set(rule.keys()) - KNOWN_RULE_KEYS
-            if unknown_keys:
-                issues.append("Rule contains unknown keys: %s." % unknown_keys)
-
-    if issues:
-        raise ValidationFailed(" ".join(issues))
-
-
-def validate_rule_port(port):
-    """
-    Validates that any value in a port list really is valid.
-    Valid values are an integer port, or a string range separated by a colon.
-
-    :param port: the port, which is validated for type
-    :returns: None or an error string if invalid
-    """
-    if isinstance(port, int):
-        if port < 1 or port > 65535:
-            return "integer out of range"
-        return None
-
-    # If not an integer, must be format N:M, i.e. a port range.
-    try:
-        fields = port.split(":")
-    except AttributeError:
-        return "neither integer nor string"
-
-    if not len(fields) == 2:
-        return "range unparseable"
-
-    try:
-        start = int(fields.pop(0))
-        end = int(fields.pop(0))
-    except ValueError:
-        return "range invalid"
-
-    if start >= end or start < 1 or end > 65535:
-        return "range invalid"
-
-    return None
 
 
 def parse_if_tags(etcd_node):
@@ -610,7 +423,7 @@ def parse_if_tags(etcd_node):
         else:
             tags = json_decoder.decode(etcd_node.value)
             try:
-                validate_tags(tags)
+                common.validate_tags(tags)
             except ValidationFailed:
                 _log.exception("Validation failed for profile %s tags : %s",
                                profile_id, tags)
@@ -620,26 +433,3 @@ def parse_if_tags(etcd_node):
 
         return profile_id, tags
     return None, None
-
-
-def validate_tags(tags):
-    """
-    Ensures that the supplied tags are valid. Once this routine has returned
-    successfully, we know that all required fields are present and have valid
-    values.
-
-    :param tags: tag set as read from etcd
-    :raises ValidationFailed
-    """
-    issues = []
-
-    if not isinstance(tags, list):
-        issues.append("Expected tags to be a list.")
-    else:
-        for tag in tags:
-            if not isinstance(tag, StringTypes):
-                issues.append("Expected tag '%s' to be a string." % tag)
-                break
-
-    if issues:
-        raise ValidationFailed(" ".join(issues))
